fix ut AudioLimiterManagerUnitTest.GetLatency_002

+++ services/audio_service/test/unittest/common/limiter_unit_test/audio_limiter_manager_unit_test.cpp
@@ -254,7 +254,7 @@ HWTEST_F(AudioLimiterManagerUnitTest, GetLatency_002, TestSize.Level1)
     ret = limiterManager->GetLatency(sinkIndex);
     EXPECT_EQ(ret,
         TEST_MAX_REQUEST * AUDIO_MS_PER_S /
-            (static_cast<uint8_t>(SAMPLE_F32LE) * static_cast<uint8_t>(SAMPLE_RATE_48000) *
+            (static_cast<uint8_t>(SAMPLE_F32LE) * static_cast<uint32_t>(SAMPLE_RATE_48000) *
                 static_cast<uint8_t>(STEREO) * PROC_COUNT));
 }
 }  // namespace AudioStandard
